+++ qupath-extension-processing/src/main/resources/qupath/imagej/gui/scripts/examples/groovy/Classify object.groovy
@@ -1,24 +1,39 @@
 package qupath.imagej.gui.scripts.examples.hidden
 /*
- * Groovy script showing how a QuPath object can be classified.
- * You will need to send the current object as a Roi, and also
- * return the active Roi to see the results in QuPath.
+ * Groovy script showing how a QuPath object can be classified
+ * by thresholding based on local intensities.
+ * 
+ * You will need to send the current object as a Roi.
+ * Because this script classifies the object directly, you should 
+ * *not* return any Roi or overlay to QuPath
+ * (since this could result in the classification being reset).
  */
 
 import ij.IJ
-import qupath.imagej.tools.IJProperties
 
 // Check we have a Roi
 var imp = IJ.getImage()
-var roi = imp == null ? null : imp.getRoi()
+var roi = imp.getRoi()
 if (!roi) {
     return
 }
 
-// Make all measurements for the Roi
-var stats = imp.getProcessor().getStatistics()
-if (stats.max - stats.mean > stats.mean - stats.min) {
-    IJProperties.setClassification(roi, "dark")
+// Determine a threshold from outside the Roi
+var roiInverse = roi.getInverse(imp)
+imp.setRoi(roiInverse)
+var statsOutside = imp.getStatistics()
+double threshold = statsOutside.mean + 3 * statsOutside.stdDev
+
+// Determine some statistics from inside the Roi
+imp.setRoi(roi)
+var statsInside = imp.getStatistics()
+
+// Get the object
+var pathObject = imp.getProperty("qupath.pathObject")
+
+// Assign a classification
+if (statsInside.mean > threshold) {
+    pathObject.classification = "Positive"
 } else {
-    IJProperties.setClassification(roi, "light")
+    pathObject.classification = "Negative"
 }